+++ packages/app/src/modules/shared/index.ts
@@ -1,3 +1,4 @@
 export { getDsfrModal } from "./getDsfrModal";
 export { PdfFileUpload } from "./PdfFileUpload";
 export { SubmitModal } from "./SubmitModal";
+export { usePdfUploadForm } from "./usePdfUploadForm";
